pyproject.toml
[tool.poetry]
name = "ml-lib"
version = "0.1.0"
description = "boilerplate for ml projects"
authors = ["Tristan BRUGÈRE <67050900+nephanth@users.noreply.github.com>"]
license = "Cecill-B"
readme = "README.md"
packages = [{include = "ml_lib"}]

[tool.poetry.dependencies]
python = "^3.11"
torch = ">=1.13.1"
sqlalchemy = "^2.0.23"
pydantic = "^2.5.3"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
seaborn = "^0.12.2"
notebook = "^7.0.4"


[tool.poetry.group.experimenttracking.dependencies]
sqlalchemy = "^2.0.10"

[tool.poetry.group.display.dependencies]
svg-py = "^1.4.2"
matplotlib = "^3.7.0"
scipy = "^1.13.0"

[tool.poetry.group.optional.dependencies]
py-trello = "^0.19.0"
tqdm = "^4.64.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
